% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdist.R
\name{calc.cdist}
\alias{calc.cdist}
\title{Calculate Daily Cumulative Distance Traveled}
\usage{
calc.cdist(moveData)
}
\arguments{
\item{moveData}{R object of class moveData from \code{moveHMM}}
}
\value{
Data.frame
}
\description{
Calculate Daily Cumulative Distance Traveled
}
