% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files2dfs.R
\name{csv2df}
\alias{csv2df}
\title{Convert Telonics Iridium CSV File}
\usage{
csv2df(
  file,
  nskip = 23,
  fixratescheds = NULL,
  QFPkeep = c("Resolved QFP", "Resolved QFP (Uncertain)"),
  addcol = NULL
)
}
\arguments{
\item{file}{Character string containing path and file name of Telonics Iridium CSV file}

\item{nskip}{Number of rows to remove from top of Iridium CSV file. This number may vary.}

\item{fixratescheds}{Optional data.frame with fix rate intervals assigned to each GPS schedule}

\item{QFPkeep}{Location quality categories to retain}

\item{addcol}{Names of columns to retain additional to default}
}
\value{
Formatted data.frame
}
\description{
Convert Telonics Iridium CSV File
}
\examples{
\dontrun{
csv2df(file = "700516A Complete.csv", nskip = 23)}
}
