% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tools_MC.R
\name{import_gps_collars}
\alias{import_gps_collars}
\title{Import Telonics GPS collar data (generalized version of `import_TelCSVfiles`)}
\usage{
import_gps_collars(
  dir_path = "./data/raw_data/telonics",
  nskip = 21,
  collar_type = "Telonics Spreadspectrum",
  ...
)
}
\arguments{
\item{dir_path}{a relative directory path to a folder containing raw Telonics GPS collar data}

\item{nskip}{a numeric value indicating the number of rows to remove from top of the CSV file}

\item{collar_type}{the make and model of GPS collar data to be imported. Accepted values include:
#'   \itemize{
    \item{"Telonics Spreadspectrum"}
    \item{"Telonics Iridium"}
    \item{"ATS Globalstar"}
    }}

\item{...}{other inherited arguments}
}
\value{
a dataframe containing GPS collar data
}
\description{
Import Telonics GPS collar data (generalized version of `import_TelCSVfiles`)
}
