% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tools.R
\name{import_tel_gps}
\alias{import_tel_gps}
\title{Import Telonics GPS collar data (generalized version of `import_TelCSVfiles`)}
\usage{
import_tel_gps(
  path,
  collar_type = NULL,
  nskip = NULL,
  report_format = NULL,
  ...
)
}
\arguments{
\item{path}{Character vector containing path(s) and file name(s) of Telonics Iridium CSV file(s),
character string specifying path to folder containing Telonics Iridium CSV file(s), or
character string specifying path to root folder containing folders with Telonics Iridium CSV file(s).\cr\cr
NOTE: When specifying a root folder within which a search of subfolders is conducted, user must pass \code{recursive = TRUE} as an additional argument.}

\item{collar_type}{Character string specifying the type of GPS collar for which data is to be imported. Accepted values include:
\itemize{
  \item{"Telonics Spreadspectrum"}
  \item{"Telonics Iridium"}
  }}

\item{nskip}{Numeric value indicating the number of rows to remove from top of the CSV file}

\item{report_format}{Character string specifying Telonics report format.  Accepted values include:
\itemize{
  \item{"Complete"}
  \item{"Condensed"}
  }}

\item{...}{Additional arguments to pass on.}
}
\value{
Dataframe containing GPS collar data.
}
\description{
Import Telonics GPS collar data (generalized version of `import_TelCSVfiles`)
}
