% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tools.R
\name{import_telcsv}
\alias{import_telcsv}
\title{Import Telonics Iridium CSV File}
\usage{
import_telcsv(file, nskip = 23, fix_attempt_keep = NULL)
}
\arguments{
\item{file}{Character string containing path and file name of Telonics Iridium CSV file}

\item{nskip}{Number of rows to remove from top of Iridium CSV file. This number may vary.}

\item{fix_attempt_keep}{Location quality categories to retain}
}
\value{
Formatted data.frame
}
\description{
Import Telonics Iridium CSV File
}
\examples{
\dontrun{
import_telcsv(file = "700516A Complete.csv", nskip = 23)}
}
