% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tools.R
\name{import_TelCSVfiles}
\alias{import_TelCSVfiles}
\title{Import Telonics Iridium CSV File(s)}
\usage{
import_TelCSVfiles(path, csv_pattern = NULL, ...)
}
\arguments{
\item{path}{Character vector containing path(s) and file name(s) of Telonics Iridium CSV file(s),
character string specifying path to folder containing Telonics Iridium CSV file(s), or
character string specifying path to root folder containing folders with Telonics Iridium CSV file(s).\cr\cr
NOTE: When specifying a root folder within which a search of subfolders is conducted, user must pass \code{recursive = TRUE} as an additional argument.}

\item{csv_pattern}{Character string specifying text pattern used to select specific CSV files.}

\item{...}{Additional arguments to pass on}
}
\value{
Formatted data.frame. NOTE: Time variables are rounded to nearest hour by default.
}
\description{
Import Telonics Iridium CSV File(s)
}
\examples{
\dontrun{

## import only using file name based current working directory
import_TelCSVfiles(path = "700516A Complete.csv")

## import using character vector specifying directory paths and file names
files <- list.files("../directory_path", pattern = "Complete", full = TRUE)
import_TelCSVfiles(path = files)

## import using character string specifying directory path to folder storing files
import_TelCSVs(path = "../directory_path", csv_pattern = "Complete")}
}
