% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tools.R
\name{import_telirid}
\alias{import_telirid}
\title{Internal Function to Import Individual Telonics Iridium CSV File}
\usage{
import_telirid(
  file,
  nskip = 23,
  fix_attempt_keep = c("Resolved QFP", "Resolved QFP (Uncertain)"),
  report_format = NULL,
  recast = TRUE,
  colname_fun = "snake_case",
  ...
)
}
\arguments{
\item{file}{Character string containing path and file name of Telonics Iridium CSV file.}

\item{nskip}{Number of rows to remove from top of Iridium CSV file. This number may vary.}

\item{fix_attempt_keep}{Character vector containing location quality categories to retain.}

\item{report_format}{Character string specifying Telonics report format of CSV files. Acceptable values are \code{Complete} and \code{Condensed}.}

\item{colname_fun}{Function used to format column names of output data.frame. Default is the internal function \code{snake_case}.}
}
\value{
Formatted data.frame.
}
\description{
Internal Function to Import Individual Telonics Iridium CSV File
}
