% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tools_MC.R
\name{import_telsst}
\alias{import_telsst}
\title{Internal Function to Import Individual Telonics Spreadspectrum CSV File (like `importtelfn`, but for Spreadspectrum collars)}
\usage{
import_telsst(
  file,
  nskip = 21,
  fix_attempt_keep = c("Succeeded (3D)", "Succeeded (2D)", "Resolved QFP",
    "Resolved QFP (Uncertain)"),
  ...
)
}
\arguments{
\item{file}{Character string containing the path and file name of a Telonics Spreadspectrum CSV file}

\item{nskip}{Number of rows to remove from top of \code{file}}

\item{fix_attempt_keep}{Character vector containing location quality categories to retain}

\item{...}{other inherited arguments}
}
\value{
Formatted data.frame
}
\description{
Internal Function to Import Individual Telonics Spreadspectrum CSV File (like `importtelfn`, but for Spreadspectrum collars)
}
