% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tools_MC.R
\name{import_telsst_complete}
\alias{import_telsst_complete}
\title{Import Telonics Spreadspectrum collar data (combined version that imports and merges CSVs)}
\usage{
import_telsst_complete(
  dir_path = NULL,
  csv_pattern = "*",
  nskip = 21,
  recursive = TRUE,
  recast = TRUE,
  fix_attempt_keep = c("Succeeded (3D)", "Succeeded (2D)", "Resolved QFP",
    "Resolved QFP (Uncertain)"),
  snake_case = FALSE,
  ...
)
}
\arguments{
\item{dir_path}{a relative directory path to a folder containing raw Telonics GPS collar data}

\item{csv_pattern}{an optional regular expression. Only file names which match the regular expression will be returned.}

\item{recursive}{a logical value. Should the listing recurse into directories? Inherited from \code{list.files()}.}

\item{recast}{a logical value. Should the variable be recast?}

\item{fix_attempt_keep}{a character vector containing location quality categories to retain. Default is:
\itemize{
  \item{"Succeeded (3D)"}
  \item{"Succeeded (2D)"}
  \item{"Resolved QFP"}
  \item{"Resolved QFP (Uncertain)"}
  }}

\item{...}{other inherited argumements}
}
\value{
A dataframe containing Telonics Spreadspectrum collar data
}
\description{
Imports, merges, and (optional) casts Telonics Spreadspectrum GPS collar data
}
\author{
McCrea Cobb \email{mccrea_cobb@fws.gov}
}
