% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tools.R
\name{import_tpf}
\alias{import_tpf}
\title{Import Telonics TPF File and Extract CTNs, IMEIs, and Fix-Rate Schedules}
\usage{
import_tpf(file)
}
\arguments{
\item{file}{Character string containing path and file name of Telonics TPF file}
}
\value{
Formatted data.frame
}
\description{
Import Telonics TPF File and Extract CTNs, IMEIs, and Fix-Rate Schedules
}
\examples{
\dontrun{
tpf2df("180906015A_1.tpf")}
}
