% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_tools.R
\name{make.moveData}
\alias{make.moveData}
\title{Convert GPS Data to moveData Class from moveHMM Package}
\usage{
make.moveData(gps_df)
}
\arguments{
\item{gps_df}{Data.frame containing GPS data following output format of csv2df function}
}
\value{
R object of class moveData
}
\description{
Convert GPS Data to moveData Class from moveHMM Package
}
\examples{
\dontrun{
gps_df <- csv2df(file = "700516A Complete.csv", nskip = 23)
make.moveData(gps_df)}
}
