% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tools.R
\name{snake_case}
\alias{snake_case}
\title{Internal Function Used to Reformat Column Names}
\usage{
snake_case(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Character vector in lower snake case format
}
\description{
Internal Function Used to Reformat Column Names
}
