% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files2dfs.R
\name{tpf2df}
\alias{tpf2df}
\title{Extract CTNs, IMEIS, and fix-rate schedules from Telonics TPF File}
\usage{
tpf2df(file)
}
\arguments{
\item{file}{Character string containing path and file name of Telonics TPF file}
}
\value{
Formatted data.frame
}
\description{
Extract CTNs, IMEIS, and fix-rate schedules from Telonics TPF File
}
\examples{
\dontrun{
tpf2df("180906015A_1.tpf")}
}
