% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdist.R
\name{cdist.fig}
\alias{cdist.fig}
\title{Write plot of daily cumulative distance traveled to PNG file}
\usage{
cdist.fig(cd, ymax, figdir.path = getwd())
}
\arguments{
\item{cd}{Data.frame containing daily cumulative distance traveled generated from \code{calc.cdist} function.}

\item{ymax}{Max limit for y axis. Scale is km.}

\item{figdir.path}{Directory path to where figure is written. Default is current working directory.}
}
\value{
PNG file written to working directory or user-specified directory path
}
\description{
Write plot of daily cumulative distance traveled to PNG file
}
\examples{
\dontrun{
telcsv <- import_telcsv(file = "700516A Complete.csv", nskip = 23)
att <- telcsv2ATT(telcsv = telcsv, animal_ids = data.frame(collar_id = "700516A", animal_id = "lynx_1"))
moveData <- make.moveData(att)
cd <- calc.cdist(moveData)
cdist.fig(cd, ymax = 4500)}
}
