% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdist.R
\name{cdist.plot}
\alias{cdist.plot}
\title{Plot Daily Cumulative Distance Traveled}
\usage{
cdist.plot(cd, ymax = 3000)
}
\arguments{
\item{cd}{Data.frame containing daily cumulative distance traveled generated from \code{calc.cdist} function.}

\item{ymax}{Max limit for y axis. Scale is km.}
}
\value{
plot
}
\description{
Plot Daily Cumulative Distance Traveled
}
\examples{
\dontrun{
gps_df <- import_telcsv(file = "700516A Complete.csv", nskip = 23)
att_df <- telcsv2ATT(telcsv = gps_df, animal_ids = data.frame(collar_id = "700516A", animal_id = "lynx_1"))
moveData <- make.moveData(att_df)
cd <- calc.cdist(moveData)
cdist.plot(cd, ymax = 4500)}
}
