% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tools.R
\name{import_telirid}
\alias{import_telirid}
\title{Internal Function to Import Individual Telonics Iridium CSV File}
\usage{
import_telirid(
  file,
  nskip = 23,
  fix_attempt_keep = c("Resolved QFP", "Resolved QFP (Uncertain)"),
  recast = TRUE,
  colname_fun = "tolower",
  ...
)
}
\arguments{
\item{file}{Character string containing path and file name of Telonics Iridium CSV file.}

\item{nskip}{Number of rows to remove from top of Iridium CSV file. This number may vary.}

\item{fix_attempt_keep}{Character vector containing location quality categories to retain.}

\item{recast}{Logical value specifying whether to recast the following selected variables:
\itemize{
  \item{"Schedule_Set"}
  \item{"Acquisition_Time"}
  \item{"Acquisition_Start_Time"}
  \item{"GPS_Fix_Time"}
  \item{"Receive_Time"}
  \item{"GPS_Fix_Attempt"}
  }}

\item{colname_fun}{Function used to format column names of output data.frame. Default is the \code{tolower} function to format as snake case.}
}
\value{
Formatted data.frame.
}
\description{
Internal Function to Import Individual Telonics Iridium CSV File
}
