% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_tools.R
\name{make.moveData}
\alias{make.moveData}
\title{Convert GPS Data to moveData Class from moveHMM Package}
\usage{
make.moveData(att, by = "animal_id")
}
\arguments{
\item{att}{Data.frame containing GPS data following output format of csv2df function}

\item{by}{Character string containing name of column to be used to partition moveHMM data. Default is set to \code{animal_id}}
}
\value{
R object of class moveData
}
\description{
Convert GPS Data to moveData Class from moveHMM Package
}
