% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_tools.R
\name{telcsv2ATT}
\alias{telcsv2ATT}
\title{Convert Telonics CSV to AniTrackTools Format}
\usage{
telcsv2ATT(telcsv = NULL, animal_ids = NULL)
}
\arguments{
\item{telcsv}{Value from \code{import_telcsv} or data.frame with column names matching those from Telonics CSV file produced by the Telonics Data Converter software.}

\item{animal_ids}{A data.frame matching CTNs (animal_id$collar_id) with animal IDs (animal_id$animal_id)}
}
\value{
A data.frame formatted for use by other functions in the \code{AniTrackTools} package
}
\description{
Convert Telonics CSV to AniTrackTools Format
}
